# api.yaml
openapi: 3.0.1
info:
  title: Eventity API
  version: v1
paths:
...
  /api/notifications:
    post:
      tags:
        - Notification
      security:
      - Bearer: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationRequestDto"
          text/json:
            schema:
              $ref: "#/components/schemas/NotificationRequestDto"
          application/*+json:
            schema:
              $ref: "#/components/schemas/NotificationRequestDto"
      responses:
        201:
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationResponseDto"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        500:
          description: Server Error
    get:
      tags:
        - Notification
      parameters:
        - name: participation_id
          in: query
          schema:
            type: string
            format: uuid
      security:
      - Bearer: []
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NotificationResponseDto"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        500:
          description: Server Error
  /api/notifications/{id}:
    get:
      tags:
        - Notification
      security:
      - Bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationResponseDto"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        500:
          description: Server Error
    delete:
      tags:
        - Notification
      security:
      - Bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: Success
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        500:
          description: Server Error
  /api/participations:
    post:
      tags:
        - Participation
      security:
      - Bearer: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParticipationRequestDto"
          text/json:
            schema:
              $ref: "#/components/schemas/ParticipationRequestDto"
          application/*+json:
            schema:
              $ref: "#/components/schemas/ParticipationRequestDto"
      responses:
        201:
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParticipationResponseDto"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        500:
          description: Server Error
    get:
      tags:
        - Participation
      security:
      - Bearer: []
      parameters:
        - name: organizer_login
          in: query
          schema:
            type: string
        - name: event_title
          in: query
          schema:
            type: string
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserParticipationInfoResponseDto"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        500:
          description: Server Error
  /api/participations/{id}:
    get:
      tags:
        - Participation
      security:
      - Bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParticipationResponseDto"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        500:
          description: Server Error
    patch:
      tags:
        - Participation
      security:
      - Bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateParticipationStatusRequestDto"
          text/json:
            schema:
              $ref: "#/components/schemas/UpdateParticipationStatusRequestDto"
          application/*+json:
            schema:
              $ref: "#/components/schemas/UpdateParticipationStatusRequestDto"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParticipationResponseDto"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        500:
          description: Server Error
    delete:
      tags:
        - Participation
      security:
      - Bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: Success
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        500:
          description: Server Error
...
components:
  schemas:
...
    ErrorResponseDto:
      type: object
      properties:
        message:
          type: string
          nullable: true
        details:
          type: string
          nullable: true
      additionalProperties: false
...
    NotificationRequestDto:
      type: object
      properties:
        participationId:
          type: string
          format: uuid
      additionalProperties: false
    NotificationResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        participationId:
          type: string
          format: uuid
        text:
          type: string
          nullable: true
        sentAt:
          type: string
          format: date-time
      additionalProperties: false
    ParticipationRequestDto:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        eventId:
          type: string
          format: uuid
        role:
          $ref: "#/components/schemas/ParticipationRoleEnum"
        status:
          $ref: "#/components/schemas/ParticipationStatusEnum"
      additionalProperties: false
    ParticipationResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        eventId:
          type: string
          format: uuid
        role:
          $ref: "#/components/schemas/ParticipationRoleEnum"
        status:
          $ref: "#/components/schemas/ParticipationStatusEnum"
      additionalProperties: false
    ParticipationRoleEnum:
      type: integer
      format: int32
      enum:
        - Participant
        - Organizer
        - Left
    ParticipationStatusEnum:
      type: integer
      enum:
        - Invited
        - Accepted
        - Rejected
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: {}
...
    UserParticipationInfoResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        dateTime:
          type: string
          format: date-time
        address:
          type: string
          nullable: true
        organizerId:
          type: string
          format: uuid
        organizerLogin:
          type: string
          nullable: true
      additionalProperties: false
...
    UpdateParticipationStatusRequestDto:
      type: object
      required:
        - status
      properties:
        status:
          $ref: "#/components/schemas/ParticipationStatusEnum"
      additionalProperties: false
...
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter JWT Bearer token **_only_**
security:
  - Bearer: []